refactor(test): split test file for error
